Uninstall script added.
Uninstall script will run if the option is set in the PMPro Core option to delete data on uninstall. We then delete the payment plan level and order meta.

payment_plans has been changed to payment_plan for the level meta key so that it's consistent with the order meta key.